refactor: command, query 패키지 분리
