+++ src/main/java/dynamicquad/agilehub/issue/controller/IssueController.java
@@ -5,8 +5,8 @@ import dynamicquad.agilehub.global.header.CommonResponse;
 import dynamicquad.agilehub.global.header.status.SuccessStatus;
 import dynamicquad.agilehub.issue.dto.IssueRequestDto;
 import dynamicquad.agilehub.issue.dto.IssueResponseDto.IssueAndSubIssueDetail;
-import dynamicquad.agilehub.issue.service.IssueQueryService;
-import dynamicquad.agilehub.issue.service.IssueService;
+import dynamicquad.agilehub.issue.service.command.IssueService;
+import dynamicquad.agilehub.issue.service.query.IssueQueryService;
 import dynamicquad.agilehub.member.dto.MemberRequestDto.AuthMember;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
