+++ src/main/java/dynamicquad/agilehub/issue/service/factory/EpicFactory.java
@@ -1,15 +1,15 @@
 package dynamicquad.agilehub.issue.service.factory;
 
 import dynamicquad.agilehub.issue.IssueType;
-import dynamicquad.agilehub.issue.domain.Issue;
-import dynamicquad.agilehub.issue.repository.IssueRepository;
 import dynamicquad.agilehub.issue.domain.Epic;
+import dynamicquad.agilehub.issue.domain.Issue;
 import dynamicquad.agilehub.issue.domain.Story;
-import dynamicquad.agilehub.issue.repository.StoryRepository;
 import dynamicquad.agilehub.issue.dto.IssueRequestDto;
 import dynamicquad.agilehub.issue.dto.IssueResponseDto;
 import dynamicquad.agilehub.issue.dto.IssueResponseDto.SubIssueDetail;
-import dynamicquad.agilehub.issue.service.ImageService;
+import dynamicquad.agilehub.issue.repository.IssueRepository;
+import dynamicquad.agilehub.issue.repository.StoryRepository;
+import dynamicquad.agilehub.issue.service.command.ImageService;
 import dynamicquad.agilehub.member.domain.Member;
 import dynamicquad.agilehub.member.dto.AssigneeDto;
 import dynamicquad.agilehub.member.service.MemberService;
